requirements.txt
biopython>=1.83
openai>=1.40.0
python-dotenv>=1.0.0
